ASoC: pxa: Clean up SND_MMP_SOC for arm64
SND_MMP_SOC selects SND_ARM, which is limited to ARM, predates ASoC
and is apparently not needed. Drop this dependency.

Limit selection of MMP_SRAM to ARM, which is a non-DT driver in
arch/arm/mach-mmp/ that is not available on ARM64.

While at it, tidy the title (Soc -> SoC).

Signed-off-by: Andreas Färber <afaerber@suse.de>