net: lora: Add Semtech SX1301
The Semtech SX1301 was the first multi-channel LoRa "concentrator".
It uses a SPI interface to the host as well as a dual SPI interface to
its radios. These two have been implemented as spi_controller, so that
the Device Tree can specify whether the respective module uses two
SX1257, two SX1255 or a combination of these or some unforeseen chipset.

This implementation is the most recent - initialization is not yet
complete, it will need to load firmware into the two on-chip MCUs.

Unfortunately there is no full datasheet with register descriptions,
only a BSD-licensed userspace HAL implementation using spidev devices.

The register and bit-field definitions are taken from the SX1301
datasheet version 2.01 dated June 2014 with the revision information
'First released version'.

The reset state and RW capability of each field is not reflected in this
patch, however from the datasheet:
"Bits and registers that are not documented are reserved. They may
include calibration values. It is important not to modify these bits and
registers. If specific bits must be changed in a register with reserved
bits, the register must be read first, specific bits modified while
masking reserved bits and then the register can be written."

Then goes on to state:
"Reserved bits should be written with their reset state, they may be
read different states."

Register caching is currently disabled.

Example for converting firmware:
  echo '#include <stdint.h>' > cal_fw.c
  echo '#define MCU_AGC_FW_BYTE 8192' >> cal_fw.c
  cat cal_fw.var >> cal_fw.c
  gcc -c cal_fw.c -o cal_fw.o
  objcopy -O binary cal_fw.o sx1301_agc_calibration.bin

v2: Updated with lora_priv -> lora_dev_priv
    Init GPIO registers during probe
    Factored out register page handling
    Cleaned up duplicate set_cs error handling
    Fixed probe error handling
    Dropped debug output
    Implemented AGC calibration
    Implemented ARB and AGC firmware loading
    Converted to devm_ based netdev registration (by Ben)
    Converted from *spi to *priv pointer passing (by Ben)
    Converted load_firmware to take firmware directly (by Ben)
    Un-duplicated firmware size checks (by Ben)
    Replaced firmware magic numbers (by Ben)
    Converted to regmap (Mark, by Ben)
    Moved firmware loading/calibration to loradev open
    Implemented regmap_bus for radio SPI (Mark, by Ben)
    Implemented radio bus and device framework
    Added error output for reset GPIO
    Adopted regmap_noinc_write() (by Ben)
    Converted to regmap fields (by Ben)
    Added radio SPI debug output
    Fixed radio SPI write
    Renamed sx1301 -> sx130x
    De-duplicated register fields array again
    Dropped unused include
    Switched regmap_bus .reg_{read,write} -> .{read,write,gather_write} (Mark)
    Switched regmap_bus regmap_{read,write}() -> regmap_raw_{read,write}()
    Factored out radio A vs. B registers for reuse
    Marked MPD register {read,write}able_noinc
    Marked radio A/B SPI data readback and AGC/ARB MCU data registers volatile
    Moved around radio child debug output

Cc: Steve deRosier <derosier@gmail.com>
Cc: Mark Brown <broonie@kernel.org>
Signed-off-by: Andreas Färber <afaerber@suse.de>
Co-Developed-by: Ben Whitten <ben.whitten@gmail.com>
Signed-off-by: Ben Whitten <ben.whitten@lairdtech.com>
Signed-off-by: Ben Whitten <ben.whitten@gmail.com>
Signed-off-by: Andreas Färber <afaerber@suse.de>

net: lora: sx130x: Clean up radio debug output

It was printing as the child that it added itself - move it to the parent.

Signed-off-by: Andreas Färber <afaerber@suse.de>